+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
 export * from './types';
 
-import createConfig from './curi';
-export default createConfig;
+import curi from './curi';
+export default curi;