+++ packages/core/types/curi.d.ts
@@ -1,5 +1,5 @@
 import { History } from '@hickory/root';
 import { RouteDescriptor } from './types/route';
-import { CuriConfig, ConfigOptions } from './types/curi';
-declare function createConfig(history: History, routeArray: Array<RouteDescriptor>, options?: ConfigOptions): CuriConfig;
-export default createConfig;
+import { CuriRouter, RouterOptions } from './types/curi';
+declare function createRouter(history: History, routeArray: Array<RouteDescriptor>, options?: RouterOptions): CuriRouter;
+export default createRouter;