+++ packages/core/types/types/curi.d.ts
@@ -3,7 +3,7 @@ import { PathFunctionOptions } from 'path-to-regexp';
 import { Addon, Addons } from './addon';
 import { RouteDescriptor } from './route';
 import { Response } from './response';
-export declare type ResponseHandler = (response: Response, action?: Action, config?: CuriConfig) => void;
+export declare type ResponseHandler = (response: Response, action?: Action, router?: CuriRouter) => void;
 export interface RespondOptions {
     once?: boolean;
 }
@@ -16,13 +16,13 @@ export interface Cache {
     set: (response: Response) => void;
     get: (location: HickoryLocation) => Response;
 }
-export interface ConfigOptions {
+export interface RouterOptions {
     addons?: Array<Addon>;
     sideEffects?: Array<SideEffect>;
     cache?: Cache;
     pathnameOptions?: PathFunctionOptions;
 }
-export interface CuriConfig {
+export interface CuriRouter {
     refresh: (routeArray: Array<RouteDescriptor>) => void;
     respond: (fn: ResponseHandler, options?: RespondOptions) => RemoveResponseHandler;
     addons: Addons;